Refactoring + edit layouts
