+++ abox/src/main/java/com/sport/abox/BoxerMovementLibraryActivity.kt
@@ -1,9 +1,10 @@
-package com.example.abox
+package com.sport.abox
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.example.abox.R
 
-class BoxerMovementLibrary : AppCompatActivity() {
+class BoxerMovementLibraryActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.boxer_movement_library)
