+++ abox/src/main/java/com/sport/abox/ItemItMenu.kt
@@ -1,0 +1,8 @@
+package com.sport.abox
+
+enum class ItemItMenu(val index: Int) {
+    SETTINGS_ID(index = 1),
+    SAVE_ID(index = 2),
+    DELETE_ALL_ID(index = 3),
+    DELETE_TRAINING_ID(index = 4),
+}
