+++ abox/src/main/java/com/sport/abox/MainActivity.kt
@@ -1,9 +1,10 @@
-package com.example.abox
+package com.sport.abox
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
+import com.example.abox.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -12,12 +13,12 @@ class MainActivity : AppCompatActivity() {
 
         val btnLibrary = findViewById<Button>(R.id.btnLibrary)
         btnLibrary.setOnClickListener {
-            startActivity(Intent(this, BoxerMovementLibrary().javaClass))
+            startActivity(Intent(this, BoxerMovementLibraryActivity::class.java))
         }
 
         val btnGenerateWorkout = findViewById<Button>(R.id.btnGenerateWorkout)
         btnGenerateWorkout.setOnClickListener {
-            startActivity(Intent(this, AllTrainingsActivity().javaClass))
+            startActivity(Intent(this, AllTrainingsActivity::class.java))
         }
     }
 }
