+++ abox/src/main/java/com/sport/abox/db/DBHelper.kt
@@ -1,4 +1,4 @@
-package com.example.abox.db
+package com.sport.abox.db
 
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
