+++ abox/src/main/res/layout/boxer_movement_library.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BoxerMovementLibrary"
+    tools:context="com.sport.abox.BoxerMovementLibraryActivity"
     >
 
     <LinearLayout
