+++ abox/src/main/res/layout/round_and_exercises.xml
@@ -1,34 +1,34 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:layout_margin="10dp">
 
-    <LinearLayout
+    <TextView
+        android:id="@+id/tvRoundOfTraining"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="vertical"
-        app:layout_constraintTop_toTopOf="parent">
+        android:gravity="center"
+        android:text=""
+        android:textColor="@color/classic_yellow"
+        android:textSize="24sp"
+        app:layout_constraintBottom_toTopOf="@+id/tvExerciseOfTraining"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
-        <TextView
-            android:id="@+id/tvRoundOfTraining"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:textSize="24sp"
-            android:textColor="@color/classic_yellow"
-            android:gravity="center"
-            android:text="" />
-
-        <TextView
-            android:id="@+id/tvExerciseOfTraining"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:textSize="16sp"
-            android:textColor="@color/white"
-            android:gravity="left"
-            android:text="" />
-    </LinearLayout>
+    <TextView
+        android:id="@+id/tvExerciseOfTraining"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:gravity="left"
+        android:text=""
+        android:textColor="@color/white"
+        android:textSize="16sp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/tvRoundOfTraining" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
