--- abox/src/main/res/layout/round_task.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/llRoundTask"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:layout_margin="10dp"
-    android:background="@color/gray_blue"
-    android:orientation="vertical"
-    >
-
-    <TextView
-        android:id="@+id/tvRoundNumber"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="top|center_horizontal"
-        android:layout_margin="5dp"
-        android:text="TextView"
-        android:textColor="#FFC107"
-        android:textSize="24sp"
-        />
-
-    <LinearLayout
-        android:id="@+id/llRoundsTask"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|center_horizontal"
-        android:layout_margin="5dp"
-        android:orientation="vertical"
-        >
-
-        <TextView
-            android:id="@+id/tvRoundTask"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="5dp"
-            android:text="TextView"
-            android:textSize="16sp"
-            />
-
-    </LinearLayout>
-
-</LinearLayout>
