+++ abox/src/main/res/layout/training_activity.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".AllTrainingsActivity">
+    tools:context="com.sport.abox.AllTrainingsActivity">
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
         android:id="@+id/btnAddRound"
@@ -19,56 +19,59 @@
         app:srcCompat="@android:drawable/ic_input_add"
         tools:ignore="MissingConstraints" />
 
-        <LinearLayout
-            android:id="@+id/ll"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            >
 
-            <Space
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_margin="10dp"
-                android:layout_weight="3" />
+    <Space
+        android:id="@+id/sp"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_margin="10dp"
+        app:layout_constraintHorizontal_weight="1"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toStartOf="@+id/tvTrainingTitle"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintBottom_toTopOf="@id/rvAllRoundsAtTraining"/>
 
-            <TextView
-                android:id="@+id/tvTrainingTitle"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_margin="10dp"
-                android:layout_weight="1"
-                android:clickable="true"
-                android:gravity="center"
-                android:text=""
-                android:textColor="#03A9F4"
-                android:textSize="30sp" />
+    <TextView
+        android:id="@+id/tvTrainingTitle"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_margin="10dp"
+        android:clickable="true"
+        android:gravity="center"
+        app:layout_constraintHorizontal_weight="5"
+        android:text=""
+        android:textColor="#03A9F4"
+        android:textSize="30sp"
+        app:layout_constraintStart_toEndOf="@+id/sp"
+        app:layout_constraintEnd_toStartOf="@+id/tvRounds"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintBottom_toTopOf="@id/rvAllRoundsAtTraining"/>
 
-            <TextView
-                android:id="@+id/tvRounds"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_margin="10dp"
-                android:layout_weight="3"
-                android:text=""
-                android:textSize="30sp" />
-        </LinearLayout>
+    <TextView
+        android:id="@+id/tvRounds"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_margin="10dp"
+        app:layout_constraintHorizontal_weight="1"
+        android:text=""
+        android:textSize="30sp"
+        app:layout_constraintStart_toEndOf="@+id/tvTrainingTitle"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintBottom_toTopOf="@id/rvAllRoundsAtTraining"/>
 
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/rvAllRoundsAtTraining"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            android:layout_marginStart="8dp"
-            android:layout_marginTop="8dp"
-            android:layout_marginEnd="8dp"
-            android:layout_marginBottom="8dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/ll"
-            tools:listitem="@layout/round_and_exercises" />
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/rvAllRoundsAtTraining"
+        android:layout_width="0dp"
+        android:layout_height="0dp"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="8dp"
+        android:layout_marginEnd="8dp"
+        android:layout_marginBottom="8dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/tvTrainingTitle"
+        tools:listitem="@layout/round_and_exercises" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
