+++ abox/src/main/res/menu/menu_workout.xml
@@ -5,35 +5,29 @@
         android:icon="@mipmap/ic_launcher_round"
         android:iconTint="@color/cardview_dark_background"
         android:iconTintMode="src_atop"
-        android:title="@string/setting"
-        >
+        android:title="@string/setting">
 
     </item>
     <item
         android:id="@+id/menu_open"
         android:icon="@android:drawable/ic_menu_set_as"
-        android:title="@string/open"
-        >
+        android:title="@string/open">
 
     </item>
     <item
         android:id="@+id/menu_save"
         android:icon="@android:drawable/ic_menu_save"
-        android:title="@string/save"
-        ></item>
+        android:title="@string/save" />
     <item
         android:id="@+id/menu_generate"
         android:icon="@android:drawable/ic_menu_sort_by_size"
-        android:title="@string/generate"
-        ></item>
+        android:title="@string/generate" />
     <item
         android:id="@+id/menu_share"
         android:icon="@android:drawable/ic_menu_share"
-        android:title="@string/share"
-        ></item>
+        android:title="@string/share" />
     <item
         android:id="@+id/menu_clear"
         android:icon="@android:drawable/ic_menu_delete"
-        android:title="@string/clear"
-        ></item>
+        android:title="@string/clear" />
 </menu>
